Move receiver class to new module
