+++ pillarbox-demo-tv/build.gradle.kts
@@ -10,6 +10,7 @@ plugins {
 dependencies {
     implementation(project(":pillarbox-core-business"))
     implementation(project(":pillarbox-core-business-cast"))
+    implementation(project(":pillarbox-cast-receiver"))
     implementation(project(":pillarbox-demo-shared"))
     implementation(project(":pillarbox-player"))
     implementation(project(":pillarbox-ui"))
